[FIX] l10n_it_split_payment: move lines are not correctly computed
When editing invoice lines, move lines are not correctly computed

Co-Author: eLBati <elbaddy@gmail.com>
Co-Author: masimassimo <dr.masi.massimo@gmail.com>
